+++ src/Cathedral/Builder/EntityAbstractBuilder.php
@@ -17,13 +17,10 @@
 namespace Cathedral\Builder;
 
 use Zend\Code\Generator\PropertyGenerator;
-use Zend\Code\Generator\PropertyValueGenerator;
-use Zend\Code\Generator\MethodGenerator;
 use Zend\Code\Generator\ParameterGenerator;
 use Zend\Code\Generator\DocBlockGenerator;
 use Zend\Code\Generator\DocBlock\Tag\ReturnTag;
 use Zend\Code\Generator\DocBlock\Tag\ParamTag;
-use Zend\Filter\Null;
 
 /**
  * Builds the Abstract Entity
@@ -70,7 +67,10 @@ class EntityAbstractBuilder extends BuilderAbstract implements BuilderInterface 
 		$getter = "get{$properyName}";
 		$setter = "set{$properyName}";
 		
-		// Extract arry to $type, $default, $primary
+		// Extract array to $type, $default, $primary
+		$type = null;
+		$default = null;
+		$primary = null;
 		extract($this->getNames()->properties[$property]);
 		
 		//METHODS
@@ -116,7 +116,7 @@ MBODY;
 		$table = substr($columnName, 3);
 		$parent = new NameManager($this->getNames()->namespace, $table);
 		// METHOD:getRelationParent
-		$method = $this->buildMethod("get{$parent->entityName}");
+		$method = $this->buildMethod("fetch{$parent->entityName}");
 		$body = <<<MBODY
 \${$parent->tableName} = new \\{$parent->namespace_model}\\{$parent->modelName}();
 return \${$parent->tableName}->get{$parent->entityName}(\$this->{$columnName});
@@ -141,10 +141,11 @@ MBODY;
 		$child = new NameManager($this->getNames()->namespace, $tableName);
 		
 		// METHOD:getRelationChild
-		$method = $this->buildMethod("get{$child->entityName}s");
+		$functionName = ucwords($tableName); 
+		$method = $this->buildMethod("gather{$functionName}");
 		$body = <<<MBODY
 \${$child->tableName} = new \\{$child->namespace_model}\\{$child->modelName}();
-return \${$child->tableName}->select(['fk_{$this->getNames()->tableName}', \$this->{$this->getNames()->primary}]);
+return \${$child->tableName}->select(['fk_{$this->getNames()->tableName}' => \$this->{$this->getNames()->primary}]);
 MBODY;
 		$method->setBody($body);
 		$tag = new ReturnTag();
@@ -171,7 +172,10 @@ MBODY;
 		$this->_class->setDocBlock($docBlock);
 		
 		foreach ($this->getNames()->properties as $name => $values) {
-			// Extract arry to $type, $default, $primary
+			// Extract array to $type, $default, $primary
+			$type = null;
+			$default = null;
+			$primary = null;
 			extract($values);
 			
 			$property = new PropertyGenerator($name);