+++ src/Cathedral/Builder/EntityBuilder.php
@@ -16,11 +16,7 @@
  
 namespace Cathedral\Builder;
 
-use Zend\Code\Generator\MethodGenerator;
-use Zend\Code\Generator\ParameterGenerator;
 use Zend\Code\Generator\DocBlockGenerator;
-use Zend\Code\Generator\DocBlock\Tag\ReturnTag;
-use Zend\Code\Generator\PropertyGenerator;
 
 /**
  * Builds the Entity